u-boot: change u-boot/install target from sentinel to directory
Rather than using a sentinenl touch file and having consumers of the u-boot
artifacts getting the files from the u-boot build directory, nominate
a selection of files to copy to the u-boot/install directory.
This is to make it simpler to copy u-boot artifacts into container images
without having to include the entire u-boot build directory in the build
context or conflicts with the autogenerated dependency.

The docs for this was previously wrong, but now happens to be more correct.