git-clone: added MAKROCOSM_GIT_CLONE_SHALLOW check to perform shallow clones
The default behaviour is to clone the full git repository into build/git
and then have the working directory be cloned from that repo. It's nice
to have the full history for development but can add a lot of time to CI
builds -- especially cloning the full history of the linux kernel.

So add a shallow clone option.