workspace: remove unused user/group
makrocosm-workspace runs the workspace container with the uid and gid
set from the host caller and gid of host docker's group (so that the
user has permission to access docker.sock still).
chmod-ing the /home directory is still important, but we don't need to
pretend there is a "ubuntu" user.